[tests] split up usersuite tests to move to `module` scoped fixtures
